feat: implement GitHub Actions CI pipeline, mock tests, and pin dependencies
