+++ .github/workflows/ci.yml
@@ -1,0 +1,41 @@
+name: B0Bot CI Pipeline
+
+on:
+  push:
+    branches: [ "main" ]
+  pull_request:
+    branches: [ "main" ]
+
+jobs:
+  build-and-test:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.9", "3.10"]
+
+    steps:
+    - name: Checkout Repository
+      uses: actions/checkout@v4
+
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v5
+      with:
+        python-version: ${{ matrix.python-version }}
+        cache: 'pip'
+
+    - name: Install Dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+        pip install pytest pytest-cov pytest-mock
+
+    - name: Create dummy .env file
+      run: |
+        echo "PINECONE_API_KEY=dummy_pinecone_key" >> .env
+        echo "PINECONE_INDEX_NAME=dummy_index" >> .env
+        echo "HUGGINGFACE_TOKEN=dummy_hf_token" >> .env
+        echo "FLASK_ENV=testing" >> .env
+
+    - name: Run Tests with Pytest
+      run: |
+        pytest tests/ -v
