+++ .gitignore
@@ -1,3 +1,4 @@
 .env
 .vercel
 __pycache__/
+.venv/
