+++ requirements.txt
@@ -1,6 +1,5 @@
 python-dotenv
 Flask
-langchain
 openai==0.27.8
 gunicorn==19.7.1
 pymongo==4.4.0
@@ -11,5 +10,8 @@ lxml==4.9.2
 httpx==0.23.3
 requests==2.28.2
 pinecone-client
-langchain-pinecone
 sentence-transformers
+langchain==0.1.16
+langchain-community==0.0.34
+langchain-core==0.1.45
+langchain-pinecone==0.1.0
